requirements.txt
flask 
flask-cors
python-dotenv
requests
langchain
faiss-cpu 
sentence-transformers 
pypdf 
pdfplumber 
transformers